+++ .flowconfig
@@ -15,8 +15,11 @@
 
 [options]
 module.name_mapper='.*\(.s?css\)' -> 'empty/object'
-module.name_mapper='^components' ->'<PROJECT_ROOT>/src/components'
-module.name_mapper='^reducers' ->'<PROJECT_ROOT>/src/redux'
+module.name_mapper='^app' ->'<PROJECT_ROOT>/src/app'
+module.name_mapper='^components' ->'<PROJECT_ROOT>/src/core/components'
+module.name_mapper='^reducers' ->'<PROJECT_ROOT>/src/core/redux'
+module.name_mapper='^utils' ->'<PROJECT_ROOT>/src/core/utils'
+module.name_mapper='^helpers' ->'<PROJECT_ROOT>/src/core/helpers'
 
 [version]
 0.59.0
