+++ src/app/components/TypicalForm/TypicalValidation.js
@@ -1,7 +1,7 @@
 // @flow
 
 import memoize from 'lru-memoize'
-import { createValidator, required, email } from '../../utils/validation'
+import { createValidator, required, email } from 'utils/validation'
 
 const typicalValidation = createValidator({
   firstName: required,
