+++ src/app/containers/Hero/Hero.scss
