+++ src/app/containers/NotFound/index.js
@@ -1,5 +1,5 @@
 import universal from 'react-universal-component'
-import Loading from 'components/Loading/Loading'
+import Loading from 'app/components/Loading/Loading'
 
 export default universal(() => import('./NotFound'), {
   loading: Loading
