+++ src/app/index.js
@@ -4,15 +4,15 @@ import React from 'react'
 import Switch from 'react-router-dom/Switch'
 import Route from 'react-router-dom/Route'
 import Helmet from 'react-helmet'
-import Menu from 'components/Menu/Menu'
-import Loading from 'components/Loading/Loading'
 import RedirectWithStatus from 'components/RouterStatus/RedirectWithStatus'
-import Home from '../Home'
-import Examples from '../Examples'
-import NotFound from '../NotFound'
-import Hero from '../Hero/Hero'
-import config from '../../config'
-import style from './App.scss'
+import Menu from 'app/components/Menu/Menu'
+import Loading from 'app/components/Loading/Loading'
+import Examples from 'app/containers/Examples'
+import Hero from 'app/containers/Hero/Hero'
+import Home from 'app/containers/Home'
+import NotFound from 'app/containers/NotFound'
+import config from '../config'
+import style from './index.scss'
 
 const App = () => (
   <div>
