+++ src/app/redux/index.js
@@ -1,0 +1,9 @@
+// @flow
+
+import neo, { fetchNeoFeedEpic } from './neo'
+
+export const appEpic = [fetchNeoFeedEpic]
+
+export default {
+  neo
+}
