+++ src/client.js
@@ -4,8 +4,8 @@ import { Provider } from 'react-redux'
 import AppContainer from 'react-hot-loader/lib/AppContainer'
 import BrowserRouter from 'react-router-dom/BrowserRouter'
 import runtime from 'serviceworker-webpack-plugin/lib/runtime'
-import configureStore from './redux/configureStore'
-import App from './containers/App/App'
+import configureStore from 'reducers/configureStore'
+import App from 'app'
 
 const supportsHistory = 'pushState' in window.history
 const { store } = configureStore(f => f, window.__data)
@@ -23,8 +23,8 @@ const renderApp = TheApp =>
   )
 
 if (module.hot) {
-  module.hot.accept('./containers/App/App.js', () => {
-    const theApp = require('./containers/App/App').default
+  module.hot.accept('app', () => {
+    const theApp = require('app').default
     renderApp(theApp)
   })
 }
