+++ src/core/redux/reducers.js
@@ -3,11 +3,11 @@
 import { combineReducers } from 'redux'
 import { combineEpics } from 'redux-observable'
 import { reducer as formReducer } from 'redux-form'
-import neo, { fetchNeoFeedEpic } from './neo'
+import reducers, { appEpic } from 'app/redux'
 
-export const rootEpic = combineEpics(fetchNeoFeedEpic)
+export const rootEpic = combineEpics(...appEpic)
 
 export default combineReducers({
-  neo,
+  ...reducers,
   form: formReducer
 })
