+++ webpack/server.dev.js
@@ -8,7 +8,7 @@ const autoprefixer = require('autoprefixer')
 const res = p => path.resolve(__dirname, p)
 
 const modeModules = res('../node_modules')
-const entry = res('../src/server/render.js')
+const entry = res('../src/core/server/render.js')
 const output = res('../build')
 
 // if you're specifying externals to leave unbundled, you need to tell Webpack
@@ -77,8 +77,11 @@ module.exports = {
   },
   resolve: {
     alias: {
-      components: path.resolve(__dirname, '..', 'src/components/'),
-      reducers: path.resolve(__dirname, '..', 'src/redux/')
+      helpers: path.resolve(__dirname, '..', 'src/core/helpers/'),
+      utils: path.resolve(__dirname, '..', 'src/core/utils/'),
+      app: path.resolve(__dirname, '..', 'src/app/'),
+      components: path.resolve(__dirname, '..', 'src/core/components/'),
+      reducers: path.resolve(__dirname, '..', 'src/core/redux/')
     },
     extensions: ['.json', '.js', '.jsx']
   },
