+++ webpack/server.prod.js
@@ -4,7 +4,7 @@ const autoprefixer = require('autoprefixer')
 
 const res = p => path.resolve(__dirname, p)
 
-const entry = res('../src/server/render.js')
+const entry = res('../src/core/server/render.js')
 const output = res('../build')
 
 module.exports = {
@@ -56,8 +56,11 @@ module.exports = {
   },
   resolve: {
     alias: {
-      components: path.resolve(__dirname, '..', 'src/components/'),
-      reducers: path.resolve(__dirname, '..', 'src/redux/')
+      helpers: path.resolve(__dirname, '..', 'src/core/helpers/'),
+      utils: path.resolve(__dirname, '..', 'src/core/utils/'),
+      app: path.resolve(__dirname, '..', 'src/app/'),
+      components: path.resolve(__dirname, '..', 'src/core/components/'),
+      reducers: path.resolve(__dirname, '..', 'src/core/redux/')
     },
     extensions: ['.json', '.js', '.jsx']
   },
